+++ src/openclaw/config.ts
@@ -3,6 +3,7 @@ import type {
   OpenClawGateway,
   OpenClawReplyListenerConfig,
 } from "./types"
+export { validateGatewayUrl } from "./gateway-url-validation"
 
 const DEFAULT_REPLY_POLL_INTERVAL_MS = 3000
 const MIN_REPLY_POLL_INTERVAL_MS = 500
@@ -96,23 +97,4 @@ export function resolveGateway(
   }
 
   return { gatewayName: mapping.gateway, gateway, instruction: mapping.instruction }
-}
-
-export function validateGatewayUrl(url: string): boolean {
-  try {
-    const parsed = new URL(url)
-    if (parsed.protocol === "https:") return true
-    if (
-      parsed.protocol === "http:" &&
-      (parsed.hostname === "localhost" ||
-        parsed.hostname === "127.0.0.1" ||
-        parsed.hostname === "::1" ||
-        parsed.hostname === "[::1]")
-    ) {
-      return true
-    }
-    return false
-  } catch {
-    return false
-  }
 }
